+++ src/components/EditSections/EditExtendedInfo/EditExtendedInfo.js
@@ -15,7 +15,6 @@ import {
 
 import { addressTypesShape } from '../../../shapes';
 
-import PasswordControl from './PasswordControl';
 import CreateResetPasswordControl from './CreateResetPasswordControl';
 import RequestPreferencesEdit from './RequestPreferencesEdit';
 
@@ -27,12 +26,9 @@ class EditExtendedInfo extends Component {
     userEmail: PropTypes.string.isRequired,
     accordionId: PropTypes.string.isRequired,
     userFirstName: PropTypes.string.isRequired,
+    username: PropTypes.string.isRequired,
     onToggle: PropTypes.func.isRequired,
   };
-
-  state = { isCredentialFieldRequired: false };
-
-  toggleCredentialFieldRequired = (_, value) => this.setState({ isCredentialFieldRequired: !!value });
 
   buildAccordionHeader = () => {
     return (
@@ -40,7 +36,7 @@ class EditExtendedInfo extends Component {
         size="large"
         tag="h3"
       >
-        {<FormattedMessage id="ui-users.extended.extendedInformation" />}
+        <FormattedMessage id="ui-users.extended.extendedInformation" />
       </Headline>
     );
   };
@@ -53,10 +49,9 @@ class EditExtendedInfo extends Component {
       userId,
       userFirstName,
       userEmail,
+      username,
       addressTypes,
     } = this.props;
-
-    const { isCredentialFieldRequired } = this.state;
 
     const accordionHeader = this.buildAccordionHeader();
     const isEditForm = !!userId;
@@ -128,27 +123,21 @@ class EditExtendedInfo extends Component {
               id="adduser_username"
               component={TextField}
               fullWidth
-              required={isCredentialFieldRequired}
               validStylesEnabled
-              onChange={this.toggleCredentialFieldRequired}
             />
           </Col>
-          {
-            isEditForm
-              ? (
-                <CreateResetPasswordControl
-                  userId={userId}
-                  email={userEmail}
-                  name={userFirstName}
-                />
-              )
-              : (
-                <PasswordControl
-                  isRequired={isCredentialFieldRequired}
-                  toggleRequired={this.toggleCredentialFieldRequired}
-                />
-              )
+          {isEditForm && username &&
+            (
+              <CreateResetPasswordControl
+                userId={userId}
+                email={userEmail}
+                name={userFirstName}
+                username={username}
+              />
+            )
           }
+        </Row>
+        <Row>
           <RequestPreferencesEdit addressTypes={addressTypes} />
         </Row>
         <br />