+++ src/settings/FeeFineSettings.js
@@ -11,11 +11,12 @@ import { Select } from '@folio/stripes/components';
 import { ControlledVocab } from '@folio/stripes/smart-components';
 import { stripesConnect, withStripes } from '@folio/stripes/core';
 
+
 import { validate } from '../components/util';
 import {
   Owners,
   CopyModal,
-  ChargeNotice
+  ChargeNotice,
 } from './FeeFinesTable';
 
 class FeeFineSettings extends React.Component {
@@ -50,6 +51,22 @@ class FeeFineSettings extends React.Component {
   static propTypes = {
     stripes: PropTypes.shape({
       connect: PropTypes.func.isRequired,
+    }).isRequired,
+    resources: PropTypes.object,
+    mutator: PropTypes.shape({
+      feefines: PropTypes.shape({
+        POST: PropTypes.func.isRequired,
+      }),
+      templates: PropTypes.shape({
+        GET: PropTypes.func.isRequired,
+      }),
+      activeRecord: PropTypes.shape({
+        update: PropTypes.func,
+      }),
+      owners: PropTypes.shape({
+        GET: PropTypes.func.isRequired,
+        PUT: PropTypes.func.isRequired,
+      }),
     }).isRequired,
     intl: intlShape.isRequired,
   };
@@ -132,6 +149,9 @@ class FeeFineSettings extends React.Component {
     const myFeeFines = feefines.filter(f => f.ownerId !== this.state.ownerId) || [];
     const label = formatMessage({ id: 'ui-users.feefines.singular' });
     const itemErrors = validate(item, index, items, 'feeFineType', label);
+    const isAutomatedFeeFineNameUsed = feefines.some(feeFine => {
+      return item.feeFineType && feeFine.automatic && item.feeFineType.toLowerCase() === feeFine.feeFineType.toLowerCase();
+    });
 
     if (Number.isNaN(Number(item.defaultAmount)) && item.defaultAmount) {
       itemErrors.defaultAmount = formatMessage({ id: 'ui-users.feefines.errors.amountNumeric' });
@@ -162,6 +182,11 @@ class FeeFineSettings extends React.Component {
           />;
       }
     }
+
+    if (isAutomatedFeeFineNameUsed) {
+      itemErrors.feeFineType = <FormattedMessage id="ui-users.feefines.errors.reservedName" />;
+    }
+
     return itemErrors;
   }
 
@@ -229,6 +254,7 @@ class FeeFineSettings extends React.Component {
 
     const preCreateHook = (item) => {
       item.ownerId = ownerId;
+      item.automatic = false;
       return item;
     };
 
@@ -249,7 +275,7 @@ class FeeFineSettings extends React.Component {
 
     return (
       <this.connectedControlledVocab
-        {...this.props}
+        stripes={this.props.stripes}
         baseUrl="feefines"
         columnMapping={{
           feeFineType: formatMessage({ id: 'ui-users.feefines.columns.type' }),
@@ -268,7 +294,7 @@ class FeeFineSettings extends React.Component {
         preCreateHook={preCreateHook}
         records="feefines"
         rowFilter={rowFilter}
-        rowFilterFunction={(item) => (item.ownerId === ownerId)}
+        rowFilterFunction={(item) => (item.ownerId === ownerId && !item.automatic)}
         sortby="feeFineType"
         validate={this.validate}
         visibleFields={['feeFineType', 'defaultAmount', 'chargeNoticeId', 'actionNoticeId']}