+++ src/views/UserDetail/UserDetail.js
@@ -20,7 +20,6 @@ import {
   expandAllFunction,
   ExpandAllButton,
   Button,
-  Icon,
   Row,
   Col,
   Headline,
@@ -28,7 +27,8 @@ import {
 } from '@folio/stripes/components';
 
 import {
-  NotesSmartAccordion
+  NotesSmartAccordion,
+  ViewCustomFieldsRecord,
 } from '@folio/stripes/smart-components';
 
 import {
@@ -72,6 +72,9 @@ class UserDetail extends React.Component {
     resources: PropTypes.shape({
       selUser: PropTypes.object,
       user: PropTypes.arrayOf(PropTypes.object),
+      addressTypes: PropTypes.shape({
+        records: PropTypes.arrayOf(PropTypes.object),
+      }),
       permissions: PropTypes.shape({
         records: PropTypes.arrayOf(PropTypes.object),
       }),
@@ -142,6 +145,7 @@ class UserDetail extends React.Component {
         permissionsSection: false,
         servicePointsSection: false,
         notesAccordion: false,
+        customFields: false,
       },
     };
   }
@@ -269,6 +273,23 @@ class UserDetail extends React.Component {
 
     return (
       <PaneMenu>
+        <IfPermission perm="ui-users.edit">
+          <FormattedMessage id="ui-users.crud.editUser">
+            {ariaLabel => (
+              <Button
+                id="clickable-edituser"
+                buttonStyle="primary"
+                style={{ visibility: !user ? 'hidden' : 'visible' }}
+                to={this.getEditLink()}
+                buttonRef={this.editButton}
+                ariaLabel={ariaLabel}
+                marginBottom0
+              >
+                <FormattedMessage id="ui-users.edit" />
+              </Button>
+            )}
+          </FormattedMessage>
+        </IfPermission>
         {
           tagsEnabled &&
           <FormattedMessage id="ui-users.showTags">
@@ -283,40 +304,9 @@ class UserDetail extends React.Component {
             )}
           </FormattedMessage>
         }
-        <IfPermission perm="ui-users.edit">
-          <FormattedMessage id="ui-users.crud.editUser">
-            {ariaLabel => (
-              <IconButton
-                icon="edit"
-                id="clickable-edituser"
-                style={{ visibility: !user ? 'hidden' : 'visible' }}
-                href={this.getEditLink()}
-                ref={this.editButton}
-                ariaLabel={ariaLabel}
-              />
-            )}
-          </FormattedMessage>
-        </IfPermission>
       </PaneMenu>
     );
   }
-
-  getActionMenu = ({ onToggle }) => {
-    return (
-      <IfPermission perm="ui-users.edit">
-        <Button
-          data-test-user-instance-edit-action
-          buttonStyle="dropdownItem"
-          onClick={onToggle}
-          to={this.getEditLink()}
-        >
-          <Icon icon="edit">
-            <FormattedMessage id="ui-users.edit" />
-          </Icon>
-        </Button>
-      </IfPermission>
-    );
-  };
 
   checkScope = () => true;
 
@@ -395,6 +385,7 @@ class UserDetail extends React.Component {
       'addressType',
       '',
     );
+    const customFields = user?.customFields || [];
 
     if (!user) {
       return (
@@ -408,7 +399,7 @@ class UserDetail extends React.Component {
       );
     } else {
       return (
-        <React.Fragment>
+        <>
           <Pane
             data-test-instance-details
             id="pane-userdetails"
@@ -422,7 +413,6 @@ class UserDetail extends React.Component {
             lastMenu={this.renderDetailMenu(user)}
             dismissible
             onClose={this.onClose}
-            actionMenu={this.getActionMenu}
           >
             <TitleManager record={getFullName(user)} />
             <Headline
@@ -484,6 +474,14 @@ class UserDetail extends React.Component {
                 addressTypes={addressTypes}
                 expanded={sections.contactInfoSection}
                 onToggle={this.handleSectionToggle}
+              />
+              <ViewCustomFieldsRecord
+                accordionId="customFields"
+                onToggle={this.handleSectionToggle}
+                expanded={sections.customFields}
+                backendModuleName="users"
+                entityType="user"
+                customFieldsValues={customFields}
               />
               <IfPermission perm="proxiesfor.collection.get">
                 <ProxyPermissions
@@ -579,7 +577,7 @@ class UserDetail extends React.Component {
             </AccordionSet>
           </Pane>
           { helperApp && <HelperApp appName={helperApp} onClose={this.closeHelperApp} /> }
-        </React.Fragment>
+        </>
       );
     }
   }