+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change history for ui-users
 
-## 2.27.0 (IN PROGRESS)
+## [3.1.0](IN PROGRESS)
+
+* Add posibility to view/create/edit custom fields on user record. UIU-1279.
+
+## [3.0.0](https://github.com/folio-org/ui-users/tree/v3.0.0) (2020-03-17)
+[Full Changelog](https://github.com/folio-org/ui-users/compare/v2.26.0...v3.0.0)
+
 * Add default settings for user status and preferred contact in new user creation. Refs UIU-1385.
 * Fix navigation paths on cancel button click on loan details, open and closed loan lists pages. Refs UIU-1377.
 * Prevent loan details opening upon click in loans action menu without selecting link. Fixes UIU-1359.
@@ -30,13 +36,22 @@
 * Enable override for renewal of declared lost items. Refs UIU-1208.
 * Add more granular call number fields to loans exports. Refs. UIU-1358.
 * Show edit button only if user has "Can edit user profile" permission. Fixes UIU-1435.
-* Send `Notify Patron` checkbox value to backend in the Confirm fee/fine cancellation modal. Refs UIU-1483. 
+* Send `Notify Patron` checkbox value to backend in the Confirm fee/fine cancellation modal. Refs UIU-1483.
 * Add UI to mark items Claim returned. Refs UIU-1213.
 * Execute loan renew in sequence. Fixes UIU-1299.
 * Fix active/inactive state for Save button. Fixes UIU-1194.
 * Display `effective call number prefix`, `call number`, `call number suffix`, `enumeration`, `chronology`, `volume` in loans contexts. Refs UIU-1347, UIU-1391.
 * Prevent change due date for declared lost items. Refs UIU-1207.
 * Tweak text for declared lost modal. Refs UIU-1444.
+* Update eslint to >= 6.2.1 or eslint-util >= 1.4.1. Refs UIU-1446.
+* Add 'Custom Fields' under User Settings to give circulation managers ability to add more fields to records. Refs UIU-1441
+* Validate presence of `user.personal` before accessing it. Fixes UIU-1510.
+* Add link to Overdue Fine Policy and Lost Item Fee Policy on Fee/Fine Details page. Refs UIU-1247.
+* Provide permissions to custom fields. Refs UIU-1521.
+* Fix bug with assign and unassign permissions to users. Refs UIU-1518.
+* Default charge and action notices not saved on manual fee/fine charge settings. Refs UIU-1486.
+* Update to `@folio/stripes` `v3.0.0` compatible version of `plugin-find-user`.
+* Happy St. Patrick's day! 🍀 🍀 🍀
 
 ## [2.26.0](https://github.com/folio-org/ui-users/tree/v2.26.0) (2019-12-05)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.25.3...v2.26.0)
@@ -61,6 +76,7 @@
 * Restore display of "invisible" permissions. Refs UIU-1372.
 * Rename "Loans: All permissions" permission to "Users: User loans view, edit, renew (all)". Refs UIU-1344.
 * Add `autoFocus` prop to `<SearchField>`. Refs UIU-1248.
+* Implement declare items lost permission. Refs UIU-1265.
 
 ## [2.25.3](https://github.com/folio-org/ui-users/tree/v2.25.3) (2019-09-26)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.25.2...v2.25.3)